move error to helper function
